# backend/public/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Train Booking API",
    "version": "1.0.0",
    "description": "Static OpenAPI spec for the Train Booking backend"
  },
  "servers": [{ "url": "/api", "description": "Local API base path" }],
  "paths": {
...
    "/route-stations": {
      "get": {
        "summary": "List route stations",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Route stations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "route_stations": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/RouteStation" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a route station",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["route_id", "station_id", "sequence"],
                "properties": {
                  "route_id": { "type": "integer", "example": 1 },
                  "station_id": { "type": "integer", "example": 2 },
                  "sequence": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route station created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Route station created successfully." },
                    "route_station": { "$ref": "#/components/schemas/RouteStation" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation errors" }
        }
      }
    },
    "/route-stations/{route_station}": {
      "get": {
        "summary": "Get a route station",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "route_station", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Route station retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "route_station": { "$ref": "#/components/schemas/RouteStation" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a route station",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "route_station", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "route_id": { "type": "integer", "example": 1 },
                  "station_id": { "type": "integer", "example": 2 },
                  "sequence": { "type": "integer", "example": 3 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route station updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Route station updated successfully." },
                    "route_station": { "$ref": "#/components/schemas/RouteStation" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation errors" }
        }
      },
      "delete": {
        "summary": "Delete a route station",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "route_station", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Route station deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Route station deleted successfully." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RouteStation": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "route_id": { "type": "integer", "example": 1 },
          "station_id": { "type": "integer", "example": 2 },
          "sequence": { "type": "integer", "example": 1 },
          "station": { "$ref": "#/components/schemas/Station" },
          "created_by": { "type": "integer", "nullable": true, "example": 1 },
          "updated_by": { "type": "integer", "nullable": true, "example": 1 },
          "deleted_by": { "type": "integer", "nullable": true, "example": null },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "example": null },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Station": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Central Station" },
          "address": { "type": "string", "example": "123 Main Street, City" },
          "is_active": { "type": "boolean", "example": true },
          "created_by": { "type": "integer", "nullable": true, "example": 1 },
          "updated_by": { "type": "integer", "nullable": true, "example": 1 },
          "deleted_by": { "type": "integer", "nullable": true, "example": null },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "example": null },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}